+++ .travis.yml
@@ -4,6 +4,6 @@ language: python
 python:
   - "3.6"
 script:
-   - pip install -e .[schema-validation]
+   - pip install -e .
    - pytest
    - openapi-spec-validator openapi.json
