+++ optimade/server/main.py
@@ -8,10 +8,9 @@ from starlette.requests import Request
 
 from .deps import EntryListingQueryParams
 from .collections import MongoCollection
-from .models import (
-    Link, Links, StructureResource, OptimadeResponseMeta, OptimadeResponseMetaQuery,
-    OptimadeStructureResponse,
-)
+from .models.jsonapi import Link, Links
+from .models.structures import StructureResource
+from. models.toplevel import OptimadeResponseMeta, OptimadeResponseMetaQuery, OptimadeStructureResponse
 
 config = ConfigParser()
 config.read(Path(__file__).resolve().parent.joinpath('config.ini'))
