+++ optimade/server/models/structures.py
@@ -1,0 +1,59 @@
+from optimade.server.models.entries import EntryResourceAttributes, EntryResource
+
+
+class StructureResourceAttributes(EntryResourceAttributes):
+    elements: str = None
+    nelements: int = None
+    chemical_formula: str = None
+    formula_prototype: str = None
+
+
+class StructureResource(EntryResource):
+    type = "structure"
+    attributes: StructureResourceAttributes
+
+
+class StructureMapper:
+    aliases = (
+        ("id", "task_id"),
+        ("local_id", "task_id"),
+        ("last_modified", "last_updated"),
+        ("formula_prototype", "formula_anonymous"),
+        ("chemical_formula", "pretty_formula"),
+    )
+
+    list_fields = (
+        "elements",
+    )
+
+    @classmethod
+    def alias_for(cls, field):
+        return dict(cls.aliases).get(field, field)
+
+    @classmethod
+    def map_back(cls, doc):
+        if "_id" in doc:
+            del doc["_id"]
+        print(doc)
+        mapping = ((real, alias) for alias, real in cls.aliases)
+        newdoc = {}
+        reals = {real for alias, real in cls.aliases}
+        for k in doc:
+            if k not in reals:
+                newdoc[k] = doc[k]
+        for real, alias in mapping:
+            if real in doc:
+                newdoc[alias] = doc[real]
+        for k in newdoc:
+            if k in cls.list_fields:
+                newdoc[k] = ",".join(sorted(newdoc[k]))
+
+        print(newdoc)
+        if "attributes" in newdoc:
+            raise Exception("Will overwrite doc field!")
+        newdoc["attributes"] = newdoc.copy()
+        newdoc["attributes"].pop("id")
+        for k in list(newdoc.keys()):
+            if k not in ("id", "attributes"):
+                del newdoc[k]
+        return newdoc
