+++ optimade/server/models/util.py
@@ -1,0 +1,5 @@
+from pydantic import ConstrainedInt
+
+
+class NonnegativeInt(ConstrainedInt):
+    ge = 0
