+++ requirements-optional.txt
@@ -1,3 +1,4 @@
 invoke==1.0.0
+openapi-spec-validator==0.2.7
 twine==1.11.0
 pytest==4.0.0
